Adicionado uma classe Abstrata base para tratamento do Sexo

+++ source/health-calc-dotnet/health-calc-pack-dotnet/Models/MacronutrienteModel.cs
@@ -8,8 +8,8 @@ namespace health_calc_pack_dotnet.Models
 {
     public class MacronutrienteModel
     {
-        public int Carboidratos { get; set; }
-        public int Proteinas { get; set; }
-        public int Gorduras { get; set; }
+        public double Carboidratos { get; set; }
+        public double Proteinas { get; set; }
+        public double Gorduras { get; set; }
     }
 }
